requirements.txt
scanpy>=1.9.0
squidpy>=1.3.0
anndata>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
torch>=2.0.0
torch-geometric>=2.4.0
tangram-sc>=1.0.4
scvi-tools>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
